+++ backend/cmd/stork-agent/main.go
@@ -2,8 +2,11 @@ package main
 
 import (
 	"fmt"
+	"net"
 	"os"
 	"os/signal"
+	"os/user"
+	"strconv"
 	"syscall"
 
 	log "github.com/sirupsen/logrus"
@@ -19,6 +22,15 @@ func runAgent(settings *cli.Context) {
 	// We need to print this statement only after we check if the only purpose is to print a version.
 	log.Printf("Starting Stork Agent, version %s, build date %s", stork.Version, stork.BuildDate)
 
+	// try register agent in the server using agent token
+	if settings.String("server-url") != "" {
+		portStr := strconv.FormatInt(settings.Int64("port"), 10)
+		if !agent.Register(settings.String("server-url"), "", settings.String("address"), portStr, false, true) {
+			fmt.Println("problem with agent registration in Stork server, exiting")
+			os.Exit(1)
+		}
+	}
+
 	// Start app monitor
 	appMonitor := agent.NewAppMonitor()
 
@@ -28,6 +40,11 @@ func runAgent(settings *cli.Context) {
 	// Prepare Prometheus exporters
 	promKeaExporter := agent.NewPromKeaExporter(settings, appMonitor)
 	promBind9Exporter := agent.NewPromBind9Exporter(settings, appMonitor)
+
+	err := storkAgent.Setup()
+	if err != nil {
+		log.Fatalf("FATAL error: %+v", err)
+	}
 
 	// Let's start the app monitor.
 	appMonitor.Start(storkAgent)
@@ -51,6 +68,40 @@ func runAgent(settings *cli.Context) {
 	c := make(chan os.Signal, 1)
 	signal.Notify(c, os.Interrupt, syscall.SIGINT)
 	<-c
+}
+
+// Helper function that checks command line options and runs registration.
+func runRegister(cfg *cli.Context) {
+	agentAddr := ""
+	agentPort := ""
+	var err error
+	if cfg.String("agent-address") != "" {
+		agentAddr, agentPort, err = net.SplitHostPort(cfg.String("agent-address"))
+		if err != nil {
+			fmt.Printf("problem with parsing agent address: %s\n", err)
+			os.Exit(1)
+		}
+	}
+
+	// check current user - it should be root or stork-agent
+	user, err := user.Current()
+	if err != nil {
+		fmt.Printf("cannot get info about current user: %s", err)
+		os.Exit(1)
+	}
+	if user.Username != "root" && user.Username != "stork-agent" {
+		fmt.Println("agent registration should be run by `root` or `stork-agent` user")
+		os.Exit(1)
+	}
+
+	// run Register
+	if agent.Register(cfg.String("server-url"), cfg.String("token"), agentAddr, agentPort, true, false) {
+		fmt.Println("registration completed successfully")
+		os.Exit(0)
+	} else {
+		fmt.Println("registration failed")
+		os.Exit(1)
+	}
 }
 
 // Prepare urfave cli app with all flags and commands defined.
@@ -139,6 +190,42 @@ func setupApp() *cli.App {
 
 			runAgent(c)
 			return nil
+		},
+		Commands: []*cli.Command{
+			{
+				Name:      "register",
+				Usage:     "register this machine in Stork server indicated by <server-url>",
+				UsageText: "stork-agent register [options]",
+				Description: `Register current agent in Stork server using provided server URL.
+
+If server access token is provided using --server-token then the agent is automatically
+authorized (server-token based registration). Otherwise, the agent requires explicit
+authorization in the server using either UI or ReST API (agent-token based registration).`,
+				Flags: []cli.Flag{
+					&cli.StringFlag{
+						Name:    "server-url",
+						Usage:   "URL of Stork server",
+						Aliases: []string{"u"},
+						EnvVars: []string{"STORK_AGENT_SERVER_URL"},
+					},
+					&cli.StringFlag{
+						Name:    "token",
+						Usage:   "access token from Stork server",
+						Aliases: []string{"t"},
+						EnvVars: []string{"STORK_AGENT_SERVER_TOKEN"},
+					},
+					&cli.StringFlag{
+						Name:    "agent-address",
+						Usage:   "address (IP or DNS) with port of current agent host, eg: 10.11.12.13:8080",
+						Aliases: []string{"a"},
+						EnvVars: []string{"STORK_AGENT_ADDRESS"},
+					},
+				},
+				Action: func(c *cli.Context) error {
+					runRegister(c)
+					return nil
+				},
+			},
 		},
 	}