+++ backend/server/restservice/middleware.go
@@ -1,21 +1,24 @@
 package restservice
 
 import (
+	"fmt"
+	"io/ioutil"
 	"net/http"
 	"os"
 	"path"
 	"strings"
+	"text/template"
 	"time"
 
 	"github.com/pkg/errors"
 	log "github.com/sirupsen/logrus"
+
 	"isc.org/stork/server/auth"
 	"isc.org/stork/server/eventcenter"
 )
 
 // Install a middleware that traces ReST calls using logrus.
 func loggingMiddleware(next http.Handler) http.Handler {
-	log.Info("installed logging middleware")
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		remoteAddr := r.RemoteAddr
 		if realIP := r.Header.Get("X-Real-IP"); realIP != "" {
@@ -44,7 +47,6 @@ func loggingMiddleware(next http.Handler) http.Handler {
 
 // Install a middleware that is serving static files for UI.
 func fileServerMiddleware(next http.Handler, staticFilesDir string) http.Handler {
-	log.Info("installed file server middleware")
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		if strings.HasPrefix(r.URL.Path, "/api") || r.URL.Path == "/swagger.json" {
 			// serve API request
@@ -64,10 +66,93 @@ func fileServerMiddleware(next http.Handler, staticFilesDir string) http.Handler
 
 // Install a middleware that is serving `server-sent events` (SSE).
 func sseMiddleware(next http.Handler, eventCenter eventcenter.EventCenter) http.Handler {
-	log.Info("installed SSE middleware")
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		if strings.HasPrefix(r.URL.Path, "/sse") {
 			eventCenter.ServeHTTP(w, r)
+		} else {
+			// pass request to another handler
+			next.ServeHTTP(w, r)
+		}
+	})
+}
+
+// Agent installer as Bash script.
+const agentInstallerScript = `#!/bin/bash
+#!/bin/bash
+set -e -x
+
+rm -f /tmp/isc-stork-agent.{deb,rpm}
+
+curl -o isc-stork-agent.deb "http://{{.ServerAddress}}{{.DebPath}}"
+
+if [ -e /etc/debian_version ]; then
+    curl -o /tmp/isc-stork-agent.deb "{{.ServerAddress}}{{.DebPath}}"
+    dpkg -i /tmp/isc-stork-agent.deb
+else
+    curl -o /tmp/isc-stork-agent.rpm "{{.ServerAddress}}{{.RpmPath}}"
+    rpm -i /tmp/isc-stork-agent.rpm
+fi
+
+systemctl start isc-stork-agent
+systemctl status isc-stork-agent
+
+su stork-agent -s /bin/sh -c 'stork-agent register -u http://{{.ServerAddress}}'
+
+`
+
+// Install a middleware that is serving Agent installer.
+func agentInstallerMiddleware(next http.Handler, staticFilesDir string) http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		if strings.HasPrefix(r.URL.Path, "/stork-install-agent.sh") {
+			pkgsDir := path.Join(staticFilesDir, "assets/pkgs")
+			files, err := ioutil.ReadDir(pkgsDir)
+			if err != nil {
+				msg := fmt.Sprintf("problem with reading '%s' directory with packages: %s\n", pkgsDir, err)
+				log.Errorf(msg)
+				w.WriteHeader(http.StatusInternalServerError)
+				fmt.Fprint(w, msg)
+				return
+			}
+
+			var debFile string
+			var rpmFile string
+			for _, f := range files {
+				if strings.HasPrefix(f.Name(), "isc-stork-agent") {
+					if strings.HasSuffix(f.Name(), ".deb") {
+						debFile = f.Name()
+					} else if strings.HasSuffix(f.Name(), ".rpm") {
+						rpmFile = f.Name()
+					}
+				}
+			}
+
+			if debFile == "" || rpmFile == "" {
+				var msg string
+				if debFile == "" {
+					msg = fmt.Sprintf("cannot file agent deb file in '%s' directory\n", pkgsDir)
+				} else {
+					msg = fmt.Sprintf("cannot file agent rpm file in '%s' directory\n", pkgsDir)
+				}
+				log.Errorf(msg)
+				w.WriteHeader(http.StatusInternalServerError)
+				fmt.Fprint(w, msg)
+				return
+			}
+
+			data := map[string]string{
+				"ServerAddress": r.Host,
+				"DebPath":       path.Join("/assets/pkgs", debFile),
+				"RpmPath":       path.Join("/assets/pkgs", rpmFile),
+			}
+			t := template.Must(template.New("script").Parse(agentInstallerScript))
+			err = t.Execute(w, data)
+			if err != nil {
+				msg := fmt.Sprintf("problem with preparing install script: %s\n", err)
+				log.Errorf(msg)
+				w.WriteHeader(http.StatusInternalServerError)
+				fmt.Fprint(w, msg)
+				return
+			}
 		} else {
 			// pass request to another handler
 			next.ServeHTTP(w, r)
@@ -80,6 +165,7 @@ func sseMiddleware(next http.Handler, eventCenter eventcenter.EventCenter) http.
 func (r *RestAPI) GlobalMiddleware(handler http.Handler, staticFilesDir string, eventCenter eventcenter.EventCenter) http.Handler {
 	// last handler is executed first for incoming request
 	handler = fileServerMiddleware(handler, staticFilesDir)
+	handler = agentInstallerMiddleware(handler, staticFilesDir)
 	handler = sseMiddleware(handler, eventCenter)
 	handler = loggingMiddleware(handler)
 	return handler