+++ backend/server/restservice/middleware_test.go
@@ -1,8 +1,11 @@
 package restservice
 
 import (
+	"io/ioutil"
 	"net/http"
 	"net/http/httptest"
+	"os"
+	"path"
 	"testing"
 
 	"github.com/stretchr/testify/require"
@@ -82,6 +85,47 @@ func TestSSEMiddleware(t *testing.T) {
 	require.False(t, requestReceived)
 
 	// let request something else than sse, it should be forwarded to nextHandler
+	req = httptest.NewRequest("GET", "http://localhost/api/users", nil)
+	w = httptest.NewRecorder()
+	handler.ServeHTTP(w, req)
+	require.True(t, requestReceived)
+}
+
+// Check if agentInstallerMiddleware works and handles requests correctly.
+func TestAgentInstallerMiddleware(t *testing.T) {
+	requestReceived := false
+	nextHandler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		requestReceived = true
+	})
+
+	tmpDir, err := ioutil.TempDir("", "mdlw")
+	require.NoError(t, err)
+	defer os.RemoveAll(tmpDir)
+	handler := agentInstallerMiddleware(nextHandler, tmpDir)
+
+	// let do some request but when there is no folder with static content
+	req := httptest.NewRequest("GET", "http://localhost/stork-install-agent.sh", nil)
+	w := httptest.NewRecorder()
+	handler.ServeHTTP(w, req)
+	resp := w.Result()
+	resp.Body.Close()
+	require.EqualValues(t, 500, resp.StatusCode)
+	require.False(t, requestReceived)
+
+	// prepare folders
+	os.Mkdir(path.Join(tmpDir, "assets"), 0755)
+	os.Mkdir(path.Join(tmpDir, "assets/pkgs"), 0755)
+
+	// let do some request
+	req = httptest.NewRequest("GET", "http://localhost/stork-install-agent.sh", nil)
+	w = httptest.NewRecorder()
+	handler.ServeHTTP(w, req)
+	resp = w.Result()
+	resp.Body.Close()
+	require.EqualValues(t, 500, resp.StatusCode)
+	require.False(t, requestReceived)
+
+	// let request something else, it should be forwarded to nextHandler
 	req = httptest.NewRequest("GET", "http://localhost/api/users", nil)
 	w = httptest.NewRecorder()
 	handler.ServeHTTP(w, req)