+++ ChangeLog.md
@@ -1,3 +1,19 @@
+* 123 [func] godfryd
+
+    Secured agent-server channel. Implemented agent deployment using
+    script downloaded from the server. The script installs deb/rpm
+    packages with stork agent. Then the script registers current
+    machine in the server performing key and certs exchange.
+    There are two ways for authorizing the agent: server-token
+    and agent-token based.
+    Added TLS to gRPC traffic between agent and server using certs
+    set up during agent registration. Added instruction on machines
+    page how to install an agent.
+    Added generating root CA and server certs, and generation of
+    server token during server startup.
+    Added API and UI for presenting and regenerating server token
+    (Gitlab #205)
+
 * 122 [func] godfryd
 
     Migrated command line processing in agent from jessevdk/go-flags